backend/requirements.txt
openai-whisper==20240930
Flask==3.1.0
gunicorn==23.0.0
google-genai==1.2.0